Change the base protocol a bit to increase packet efficiency.
Instead of sending important packets seperately they are now send in a single buffered stream.
In a small test this increased packet size from an average of ~500-600 bytes up to around 1.4 kB.
This also reduces the number of packets that need to be waited on and thus should decrease the average wait time until a packet is ready for usage.

But the main reason this was implemented is for block updates which consist of only ~16 bytes but had a 34 byte header overhead. With the new system the overhead is only around 2-3 bytes.